change constructor to Dimension2D

+++ src/main/java/it/unibo/pensilina14/bullet/ballet/common/Dimension2Dimpl.java
@@ -7,7 +7,7 @@ public class Dimension2Dimpl implements Dimension2D {
     private final double height;
     private final double width;
       
-    public Dimension2Dimpl(final int height, final int width) {
+    public Dimension2Dimpl(final double height, final double width) {
       this.height = height;
       this.width = width;
     }
